Various fixes to Zend\Tool exception refactorings & supporting tests
