+++ library/Zend/Tool/Project/Provider/Exception/RuntimeException.php
@@ -3,7 +3,7 @@
 namespace Zend\Tool\Project\Exception;
 
 class RuntimeException
-    extends \RuntimeException
+    extends \Zend\Tool\Project\Exception\RuntimeException
     implements \Zend\Tool\Project\Exception
 {
 }
