--- tests/Zend/Tool/Project/_files/project1/.zfproject.xml
@@ -1,48 +1,0 @@
-<?xml version="1.0"?>
-<projectProfile>
-  <projectDirectory>
-    <projectProfileFile/>
-    <applicationDirectory>
-      <apisDirectory enabled="false"/>
-      <configsDirectory>
-          <applicationConfigFile type="ini" />
-      </configsDirectory>
-      <controllersDirectory>
-        <controllerFile controllerName="index"/>
-        <controllerFile controllerName="error"/>
-      </controllersDirectory>
-      <layoutsDirectory enabled="false"/>
-      <modelsDirectory/>
-      <modulesDirectory enabled="false"/>
-      <viewsDirectory>
-        <viewScriptsDirectory>
-          <viewControllerScriptsDirectory forControllerName="index">
-            <viewScriptFile scriptName="index"/>
-          </viewControllerScriptsDirectory>
-        </viewScriptsDirectory>
-        <viewHelpersDirectory/>
-        <viewFiltersDirectory enabled="false"/>
-      </viewsDirectory>
-      <bootstrapFile/>
-    </applicationDirectory>
-    <dataDirectory enabled="false">
-      <cacheDirectory enabled="false"/>
-      <searchIndexesDirectory enabled="false"/>
-      <localesDirectory enabled="false"/>
-      <logsDirectory enabled="false"/>
-      <sessionsDirectory enabled="false"/>
-      <uploadsDirectory enabled="false"/>
-    </dataDirectory>
-    <libraryDirectory>
-      <zfStandardLibraryDirectory/>
-    </libraryDirectory>
-    <publicDirectory>
-      <publicStylesheetsDirectory enabled="false"/>
-      <publicScriptsDirectory enabled="false"/>
-      <publicImagesDirectory enabled="false"/>
-      <publicIndexFile/>
-      <htaccessFile/>
-    </publicDirectory>
-    <projectProvidersDirectory enabled="false"/>
-  </projectDirectory>
-</projectProfile>
